Fixed parameters to UserActivityWorkflow..
